Task solved. Need review last question

+++ t4/src/br/ufsc/ine5431/praticaiv/PSNR.java
@@ -34,22 +34,22 @@ public final class PSNR {
 			final double result = Math.pow((Math.pow(2, bpp) - 1), 2);
 			return 10 * Math.log10(result/mse);
 	  }
-	 
+
 	 private static double mse(int[][][] original, int[][][] decodificado)  {
 		 /* TODO
 		  * Implemente aqui o cálculo do MSE. Dica: não esqueça de aplicar o cast (double) e divisões de números inteiros
 		  */
 		 int nlinhas = original[0].length;
 		 int ncolunas = original.length;
-		 int differencesSum = 0;
+		 double differencesSum = 0;
 		 for (int i=0;i<nlinhas;i++) {  //percorre linhas
 			 for (int j=0;j<ncolunas;j++) {  //percorre colunas
 				for (int p=0;p<3;p++) {  //percorre componentes R, G e B
 						differencesSum += Math.pow((original[i][j][p] - decodificado[i][j][p]), 2);
-				} 
+				}
 			 }
-		 } 
-		 return differencesSum / (nlinhas * ncolunas * 3);	 
+		 }
+		 return differencesSum / (nlinhas * ncolunas * 3);
 	}
 
 }
